Animation and camera changes, frametime fixes
Camera now scales with the pixel multiplier (creating a GameWindow with
2x pixel multiplier creates a 2x sized camera)

Animations are now independent from tilesets (previous .animation files
need to be fixed)

Settings.Frametime.StaticValue is now a float value (from int)

+++ Data/Settings.cs
@@ -39,7 +39,7 @@ namespace SFMLStart.Data
         #region Nested type: Frametime
         public static class Frametime
         {
-            public static int StaticValue = 1;
+            public static float StaticValue = 1;
             public static bool IsStatic = false;
         }
         #endregion
